requirements.txt
streamlit
PyPDF2
python-docx
camelot-py[cv]
pdfplumber
pandas
pytesseract
python-dotenv
langchain
langchain_openai
pdf2image